Split status_collector in its own file
